requirements.txt
annotated-types==0.6.0
anyio==4.3.0
black==24.4.0
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.2
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.7.0
pydantic_core==2.18.1
pyflakes==3.2.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
